requirements.txt
aiohttp  
aiogram  
Django  
django-extensions  
django-jazzmin  
django-modeltranslation  
django-rosetta  
django-unfold  
djangorestframework  
drf-material  
environs  
google-api-python-client  
google-auth-httplib2  
google-auth-oauthlib  
gspread  
gspread-formatting  
icecream  
oauth2client  
openpyxl  
python-dotenv  
requests  
uvicorn